multiple code improvements: squid:S1155, squid:S1905, squid:S1153

+++ isoparser/src/main/java/org/mp4parser/boxes/iso14496/part12/SampleSizeBox.java
@@ -91,7 +91,7 @@ public class SampleSizeBox extends AbstractFullBox {
         sampleCount = CastUtils.l2i(IsoTypeReader.readUInt32(content));
 
         if (sampleSize == 0) {
-            sampleSizes = new long[(int) sampleCount];
+            sampleSizes = new long[sampleCount];
 
             for (int i = 0; i < sampleCount; i++) {
                 sampleSizes[i] = IsoTypeReader.readUInt32(content);
